fix: :bug: Fixed not existing assets bug
Spawned the dragged asset when it doesn't exist in the drop account. Despawn asset from sending account. Moved all drag and drop logic to a plugin
